// waspc/packages/studio/client/src/appSpec.ts
export type Decl = {
  [Type in keyof DeclTypeToValue]: {
    declType: Type;
    declName: string;
    declValue: DeclTypeToValue[Type];
  };
}[keyof DeclTypeToValue];

export type DeclTypeToValue = {
  App: App;
  Page: Page;
  Route: Route;
  Query: Query;
  Action: Action;
  Job: Job;
  Api: Api;
  ApiNamespace: ApiNamespace;
  Crud: Crud;
  Entity: Entity;
};

export type GetDeclForType<T extends Decl["declType"]> = Extract<
  Decl,
  { declType: T }
>;

export type DeclType = Decl["declType"];

export type Page = {
  component: ExtImport;
  authRequired: Optional<boolean>;
};

export type Route = {
  path: string;
  to: Ref<"Page">;
};

export type Action = {
  fn: ExtImport;
  entities: Optional<Ref<"Entity">[]>;
  auth: Optional<boolean>;
};

export type Query = {
  fn: ExtImport;
  entities: Optional<Ref<"Entity">[]>;
  auth: Optional<boolean>;
};

export type Job = {
  executor: JobExecutor;
  perform: Perform;
  schedule: Optional<Schedule>;
  entities: Optional<Ref<"Entity">[]>;
};
export type Schedule = {
  cron: string;
  args: Optional<object>;
  executorOptions: Optional<ExecutorOptions>;
};

export type Perform = {
  fn: ExtImport;
  executorOptions: Optional<ExecutorOptions>;
};

export type Api = {
  fn: ExtImport;
  middlewareConfigFn: Optional<ExtImport>;
  entities: Optional<Ref<"Entity">[]>;
  httpRoute: HttpRoute;
  auth: Optional<boolean>;
};

export type ApiNamespace = {
  middlewareConfigFn: ExtImport;
  path: string;
};

export type Crud = {
  entity: Ref<"Entity">;
  operations: CrudOperations;
};

export type App = {
  wasp: Wasp;
  title: string;
  head: Optional<string[]>;
  auth: Optional<Auth>;
  server: Optional<Server>;
  client: Optional<Client>;
  db: Optional<Db>;
  emailSender: Optional<EmailSender>;
  webSocket: Optional<WebSocket>;
};
// ...
export type Ref<T extends DeclType> = {
  name: string;
  declType: T;
};

export type AnyRef = {
  [Type in keyof DeclTypeToValue]: Ref<Type>
}[keyof DeclTypeToValue];
// ...
export type Entity = Record<string, any>;